requirement.txt
flask_monitoringdashboard
flask_migrate
flask_script
flask_bcrypt
flask_testing
flask_restplus
flask-cache
flask_bootstrap
flask_wtf
jwt
pyjwt
gensim
pandas
konlpy
tqdm
celery
networkx
community
regex
beautifulsoup4==4.6.0
scikit-learn
matplotlib
PyHamcrest
pyLDAvis
python-louvain
Flask-Threads
werkzeug==0.16.0
billiard 

#conda install pytorch torchvision cpuonly -c pytorch
